pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "face_recognition"
dynamic = ["version"]
description = "This project focuses on developing a Face Recognition System for identifying Bollywood celebrities using deep learning techniques."
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Sudhanshu Rastogi", email = "sudhanshu2198@gmail.com" },
]

dependencies = ["dlib==19.22.99",
                "fastapi==0.115.10",
                "mkdocs==1.6.1",
                "mkdocs-material==9.6.6",
                "mkdocs-material-extensions==1.3.1",
                "numpy==1.26.0",
                "onnx==1.17.0",
                "onnxruntime==1.19.2",
                "opencv-python==4.11.0.86",
                "pillow==11.1.0",
                "python-multipart==0.0.20",
                "streamlit==1.42.2",
                "uvicorn==0.34.0"]

[project.urls]
Homepage = "https://github.com/sudhanshu2198/Face-Recognition-System"
Documentation = "https://sudhanshu2198.github.io/Face-Recognition-System/"
Streamlit = "https://github.com/fastapi/fastapi"
FastAPI = "https://github.com/fastapi/fastapi/issues"

[project.optional-dependencies]
dev = ["ruff","pre-commit"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4
target-version = "py39"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F","B","C901"]
fixable = ["ALL"]
unfixable = ["B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"
